+++ source/Analysis/AnnualRecurrenceProbability_Dialog.vb
@@ -1,0 +1,37 @@
+'BlueM.Wave
+'Copyright (C) BlueM Dev Group
+'<https://www.bluemodel.org>
+'
+'This program is free software: you can redistribute it and/or modify
+'it under the terms of the GNU Lesser General Public License as published by
+'the Free Software Foundation, either version 3 of the License, or
+'(at your option) any later version.
+'
+'This program is distributed in the hope that it will be useful,
+'but WITHOUT ANY WARRANTY; without even the implied warranty of
+'MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+'GNU Lesser General Public License for more details.
+'
+'You should have received a copy of the GNU Lesser General Public License
+'along with this program.  If not, see <https://www.gnu.org/licenses/>.
+'
+Friend Class AnnualRecurrenceProbability_Dialog
+
+    Public Sub New()
+
+        ' This call is required by the Windows Form Designer.
+        InitializeComponent()
+
+        ' Add any initialization after the InitializeComponent() call.
+
+        Me.ComboBox_startMonth.BeginUpdate()
+        Me.ComboBox_startMonth.Items.AddRange(Helpers.CalendarMonths.ToArray)
+        Me.ComboBox_startMonth.EndUpdate()
+
+        Me.ComboBox_startMonth.SelectedIndex = 10 'November
+    End Sub
+
+    Private Sub AnnualRecurrenceProbability_Dialog_Load(sender As Object, e As EventArgs) Handles MyBase.Load
+
+    End Sub
+End Class